requirements.txt
pyqt5
pymupdf
pillow